renaming and restructuing backend
