+++ server/src/service/index.ts
@@ -1,0 +1,6 @@
+import * as authService from './authService';
+import * as classroomService from './classroomService';
+import * as mathService from './math';
+import * as userService from './userService';
+
+export { authService, mathService, userService, classroomService };
