+++ server/src/service/math/generate/generateNumbers.ts
