+++ server/src/service/math/generate/index.ts
