--- server/src/services/index.ts
@@ -1,6 +1,0 @@
-import * as authService from './auth';
-import * as classroomService from './classroom';
-import * as mathService from './math';
-import * as userService from './user';
-
-export { authService, mathService, userService, classroomService };
