+++ server/src/web/mathRoutes.ts
@@ -1,8 +1,8 @@
 import express from 'express';
 
 import { ProblemTemplateModel } from '@server/database';
-import { mathService } from '@server/services';
-import { enhanceHandler, HTTP_CODE } from '@server/utils/routeEnhancer';
+import { mathService } from '@server/service';
+import { enhanceHandler, HTTP_CODE } from '@server/service/utils/routeEnhancer';
 
 export const initializeMathRoutes = (app: express.Application) => {
     const mathRouter = express.Router();
